Actually have different room data now  - Copied the data before, but pointers meant it was still using the same data. Rookie mistake :P
Changed MapManager to take in the actual text data and store that in RoomTemplate.
When making a MapRoom, new tiles are created and then stored there themselves from the data from the RoomTemplate files.

Also added in a main menu state with some very basic buttons that transport the user to either the gamestate or EditorState.